# arcx_backend/openapi.yaml
openapi: 3.0.3
info:
  title: ARCX Platform API
  version: 1.0.0
  description: |-
    ## ARCX — Algorithmic Reserve Currency Exchange

    ARCX is an INR-denominated, NAV-backed digital asset platform. Investors deposit INR which is converted to ARCX tokens at the current NAV price. The vault holds a diversified portfolio (SPY, TLT, GLD) rebalanced algorithmically.

    ### Authentication
    All protected endpoints require a **Bearer JWT** in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```

    Obtain tokens via `POST /api/auth/token/` or `POST /api/v1/auth/register`.

    ### Idempotency
    Financial mutation endpoints (deposit, withdraw, transfer) require an `Idempotency-Key` UUID header to prevent duplicate transactions.

    ### Error Format
    All errors follow: `{"error": "message", "code": "SNAKE_CASE_CODE"}`
  contact:
    name: ARCX Engineering
    email: engineering@arcx.in
  license:
    name: Proprietary
paths:
  /api/v1/auth/register:
    post:
      operationId: auth_register
      description: |-
        Creates a new ARCX user account and immediately returns JWT tokens so the user is logged in on the same request — no separate login step required.

        **KYC Note:** New accounts start with `kyc_status=pending`. Deposit, withdraw, and transfer all require KYC approval.
      summary: Register a new account
      tags:
      - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestRequest'
            examples:
              TypicalRegistration:
                value:
                  email: rahul@example.com
                  password: MyStr0ng#Pass
                  full_name: Rahul Sharma
                  phone: +91 98765 43210
                summary: Typical registration
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequestRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                EmailConflict:
                  value:
                    error: An account with this email already exists.
                    code: REGISTRATION_FAILED
                  summary: Email conflict
          description: Email already registered
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error — check field format
  /api/v1/auth/me:
    get:
      operationId: auth_me
      description: Returns the authenticated user's full profile, including wallet
        balance and KYC tier. One call gives everything the dashboard home screen
        needs.
      summary: Get current user profile
      tags:
      - Auth
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing or invalid JWT
  /api/v1/auth/logout:
    post:
      operationId: auth_logout
      description: Blacklists the provided refresh token so it can't mint new access
        tokens. The current access token still works until its 30-minute TTL expires
        — this is by design with stateless JWT architecture.
      summary: Logout (blacklist refresh token)
      tags:
      - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequestRequest'
            examples:
              LogoutRequest:
                value:
                  refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                summary: Logout request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LogoutRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LogoutRequestRequest'
        required: true
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid or already-blacklisted token
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing or invalid JWT
...
  /api/v1/kyc/status:
    get:
      operationId: kyc_status
      description: Returns the user's current overall KYC status, the highest approved
        tier, the resulting daily transaction limit, and the full list of all KYC
        submissions.
      summary: Get current KYC status and history
      tags:
      - KYC
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCStatusResponse'
              examples:
                KYCStatusResponse:
                  value:
                    kyc_status: approved
                    highest_tier: tier_1
                    daily_limit_inr: Rs.10,000
                    records:
                    - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                      tier: tier_1
                      status: approved
                      document_type: aadhaar
                      verified_at: '2025-06-01T15:30:00Z'
                      submitted_at: '2025-06-01T15:29:50Z'
                  summary: KYC status response
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCErrorResponse'
          description: JWT missing or expired
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable snake_case error code
      required:
      - code
      - error
...
    KYCErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
      - code
      - error
    KYCStatusResponse:
      type: object
      properties:
        kyc_status:
          type: string
          description: Overall KYC status on the user account
        highest_tier:
          type: string
          nullable: true
          description: Highest approved KYC tier
        daily_limit_inr:
          type: string
          description: Current daily transaction limit based on KYC tier
        records:
          type: array
          items:
            type: object
            additionalProperties: {}
            description: 'KYC record: id, tier, status, document_type, verified_at,
              submitted_at'
          description: All KYC submissions for this user
      required:
      - daily_limit_inr
      - highest_tier
      - kyc_status
      - records
...
    KycStatusEnum:
      enum:
      - pending
      - approved
      - rejected
      - suspended
      type: string
      description: |-
        * `pending` - Pending
        * `approved` - Approved
        * `rejected` - Rejected
        * `suspended` - Suspended
    LogoutRequestRequest:
      type: object
      description: |-
        POST /api/v1/auth/logout

        Takes the refresh token. Blacklists it so it can't be used to
        generate new access tokens. The current access token expires
        naturally (30 min) — there's no server-side access token invalidation
        in the simplejwt default setup.
      properties:
        refresh_token:
          type: string
          minLength: 1
      required:
      - refresh_token
    LogoutSuccess:
      type: object
      properties:
        message:
          type: string
      required:
      - message
...
    RegisterRequestRequest:
      type: object
      description: |-
        Validates new user registration.

        password:  Must be at least 8 characters. We don't enforce complexity
                   rules here (too annoying for demos) but production should.
        phone:     Optional at registration. Can be added during Tier 1 KYC.
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 8
        full_name:
          type: string
          minLength: 1
          maxLength: 255
        phone:
          type: string
          maxLength: 20
      required:
      - email
      - full_name
      - password
    RegisterResponse:
      type: object
      description: |-
        What we return after successful registration.
        We return JWT tokens immediately so the user is logged in right away.
        No separate login step required.
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        kyc_status:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
        message:
          type: string
      required:
      - access_token
      - email
      - full_name
      - kyc_status
      - message
      - refresh_token
      - user_id
...
    UserProfile:
      type: object
      description: |-
        GET /api/v1/auth/me — full profile with wallet summary.
        The wallet balance is nested in the response — one API call gets everything.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          maxLength: 254
        full_name:
          type: string
          maxLength: 255
        phone:
          type: string
          nullable: true
          maxLength: 20
        kyc_status:
          $ref: '#/components/schemas/KycStatusEnum'
        is_active:
          type: boolean
        wallet_id:
          type: string
          format: uuid
          readOnly: true
        arcx_balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,18})?$
          readOnly: true
        cost_basis_inr:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,4})?$
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - arcx_balance
      - cost_basis_inr
      - created_at
      - email
      - full_name
      - id
      - wallet_id
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...